// weatherradar/weatherradar-web/src/app/store/type/weather.action.type.ts
export enum WeatherActionsType {
  LOAD_WEATHER = '[Weather] Load All Weather',
  LOAD_WEATHER_SUCCESS = '[Weather] Load All Weather Success',
  LOAD_WEATHER_FAILURE = '[Weather] Load All Weather Failure',
  UPDATE_WEATHER = '[Weather] Update Weather',
  UPDATE_WEATHER_SUCCESS = '[Weather] Update Weather Success',
  UPDATE_WEATHER_FAILURE = '[Weather] Update Weather Failure',
  START_LIVE_WEATHER_FEED = '[Weather] Start Live Weather Feed',
  STOP_LIVE_WEATHER_FEED = '[Weather] Stop Live Weather Feed'
}